Extract abstract ability base class
