+++ ui/app/abilities/client.js
@@ -1,17 +1,11 @@
-import { Ability } from 'ember-can';
-import { inject as service } from '@ember/service';
+import AbstractAbility from './abstract';
 import { computed, get } from '@ember/object';
-import { equal, or, not } from '@ember/object/computed';
+import { or } from '@ember/object/computed';
 
-export default Ability.extend({
-  token: service(),
-
+export default AbstractAbility.extend({
   // Map abilities to policy options (which are coarse for nodes)
   // instead of specific behaviors.
   canWrite: or('bypassAuthorization', 'selfTokenIsManagement', 'policiesIncludeNodeWrite'),
-
-  bypassAuthorization: not('token.aclEnabled'),
-  selfTokenIsManagement: equal('token.selfToken.type', 'management'),
 
   policiesIncludeNodeWrite: computed('token.selfTokenPolicies.[]', function() {
     // For each policy record, extract the Node policy
